Allow for dependencies that are yet to be defined
Make use of secondary expansion so that you can refer in your
dependencies to variables that will be defined in Rules.mk that will be
read later.  If you happen to have such a need then simply add such
variable but delay its expansion by quoting the $ sign.

I have also restricted type of deps that go to AR to be only objects.
I just don't find a good use case for not doing so :P